Add Additional Print Columns
Add additional print columns the the NamespaceRole and
NamespaceRoleBinding CRDs, to show additional information.

For a NamespaceRole we are now showing the list of namespaces and the
selector to get all created ClusterRoles / Roles.

For a NamespaceRoleBinding we are now showing the referenced
NamespaceRole and the selector to get all ClusterRoleBindings /
RoleBindings.